QUED-37 adds new navigation level to project
