+++ .editorconfig
@@ -1,0 +1,5 @@
+# editorconfig.org
+root = true
+
+[*]
+indent_size = 2
