+++ edition-archive-frontend/assets/scss/_project.scss
@@ -1,5 +1,5 @@
 // basic layout
-body {
+.body-box {
   display: flex;
   flex-direction: column;
   height: 100vh;
