+++ edition-archive-frontend/pages/[[path]].vue
@@ -5,28 +5,28 @@
     </div>
   </div>
 </template>
+
 <script>
-export default {}
+  export default {}
 </script>
+
 <script setup>
-const {path} = useRoute()
-import {createError} from 'h3'
-
-
-const {data, error} = await useAsyncData(`content-${path}`, () => {
-  const p1 = path.lastIndexOf("/") === path.length - 1 ? path + "main" : path;
-  return queryContent().where({_path: p1}).findOne()
-})
-if (error.value) {
-  throwError(
-      createError({
-        statusCode: 404,
-        statusMessage: 'Not Found',
-      })
-  );
-}
-
+  const {path} = useRoute()
+  import {createError} from 'h3'
+  const {data, error} = await useAsyncData(`content-${path}`, () => {
+    const p1 = path.lastIndexOf("/") === path.length - 1 ? path + "main" : path;
+    return queryContent().where({_path: p1}).findOne()
+  })
+  if (error.value) {
+    throwError(
+        createError({
+          statusCode: 404,
+          statusMessage: 'Not Found',
+        })
+    );
+  }
 </script>
+
 <style scoped>
 
 </style>
