+++ edition-archive-frontend/pages/gallia_pontifica_online/publikationen.vue
@@ -1,0 +1,33 @@
+<template>
+  <GalliaPontificaOnlineLayout>
+
+    <template #content>
+        <ContentRenderer v-if="data" :value="data"/>
+    </template>
+
+  </GalliaPontificaOnlineLayout>
+</template>
+<script>
+export default {}
+</script>
+<script setup>
+const {path} = useRoute()
+import {createError} from 'h3'
+
+const {data, error} = await useAsyncData(`content-${path}`, () => {
+  const p1 = path.lastIndexOf("/") === path.length - 1 ? path + "publikationen" : path;
+  return queryContent().where({_path: p1}).findOne()
+})
+if (error.value) {
+  throwError(
+      createError({
+        statusCode: 404,
+        statusMessage: 'Not Found',
+      })
+  );
+}
+
+</script>
+<style scoped>
+
+</style>
