+++ edition-archive-frontend/pages/index.vue
@@ -30,9 +30,7 @@
 
   </div>
 </template>
+
 <script>
-
-export default {
-
-}
+  export default {}
 </script>
