+++ edition-archive-frontend/plugins/useBootstrap.client.ts
@@ -1,0 +1,5 @@
+import bootstrap from 'bootstrap/dist/js/bootstrap.bundle'
+
+export default defineNuxtPlugin(nuxtApp => {
+    nuxtApp.provide('bootstrap', bootstrap)
+})
